Support compressed content encodings
Curl supports several HTTP content encodings (such as deflate and gzip)
for compressing responses.  Pass the --compressed option to enable
advertising support to servers via the Accept-Encoding HTTP header.

This won't have much of an effect for compressed audio if servers choose
to compress it, but for the XML it is significant.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>